# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Go 语言构建的轻量级、高性能通用电商后端系统。",
        "title": "GoShop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3233",
    "basePath": "/",
    "paths": {
...
        "/api/admin/orders/{id}/refund/audit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aftersale"
                ],
                "summary": "审核售后退款",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审核参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AuditRefundReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/checkout/preview": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "结算试算",
                "parameters": [
                    {
                        "description": "试算参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/checkout.PreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/checkout.Preview"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "查询订单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "订单状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "下单参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "查询订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.Detail"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/refund": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aftersale"
                ],
                "summary": "申请售后退款",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "退款参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RefundReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/pay": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "模拟支付订单",
                "parameters": [
                    {
                        "description": "支付参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PayReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/payments": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "创建支付单",
                "parameters": [
                    {
                        "description": "支付单参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreatePaymentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/payments/callback/mock": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "模拟支付回调",
                "parameters": [
                    {
                        "description": "回调参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.MockPaymentCallbackReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "查询支付单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "支付单 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaymentOrder"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "checkout.PreviewRequest": {
            "type": "object",
            "required": [
                "addressId",
                "items"
            ],
            "properties": {
                "addressId": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/checkout.ItemReq"
                    }
                },
                "userCouponId": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.AuditRefundReq": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"approve\" 或 \"reject\"",
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateOrderReq": {
            "type": "object",
            "required": [
                "addressId",
                "items"
            ],
            "properties": {
                "addressId": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.OrderItemReq"
                    }
                },
                "userCouponId": {
                    "description": "可选",
                    "type": "integer"
                }
            }
        },
...
        "handlers.PayReq": {
            "type": "object",
            "required": [
                "orderId"
            ],
            "properties": {
                "orderId": {
                    "type": "string"
                }
            }
        },
...
    }
}